sesameBackendPy/requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
PyMySQL==1.1.1